#6 Allow navigating back from vaults.
