+++ lib/widget/entry/timed_password_widget.dart
@@ -27,6 +27,7 @@ class _TimedPassword extends State<TimedPasswordWidget> {
   }
 
   void resetState() {
+    if (!this.mounted) return;
     setState(() {
       this.password = this.widget.entry.genPassword();
     });
